pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "ddgs>=9.2.3",
    #    "duckduckgo-search>=8.1.1",
    "duckduckgo-search>=8.1.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rich>=14.0.0",
]
